Fix console support with cells
The (relatively recent) security fixes for consoleauth do not work
with cells because consoleauth and the compute manager for the instance
talk to different rabbit servers.  consoleauth is in the API cell and
the compute manager is in a child cell.

This patch adds code to proxy the following via cells:

1) compute -> consoleauth's delete_tokens_for_instance()
2) consoleauth -> compute's validate_console_port()

Also: On instance deletion when the compute manager tells consoleauth to
delete tokens for the instance, it's not necessary to do an rpc.call and
block.  The only purpose it could serve would be log a traceback on failure
on the compute manager side, but that seems unnecessary.  There's no
return value.  This patch changes it to an rpc.cast instead.

Fixes bug 1154755

+++ nova/compute/manager.py
@@ -40,6 +40,7 @@ from eventlet import greenthread
 from oslo.config import cfg
 
 from nova import block_device
+from nova.cells import rpcapi as cells_rpcapi
 from nova.cloudpipe import pipelib
 from nova import compute
 from nova.compute import instance_types
@@ -176,6 +177,7 @@ CONF.import_opt('host', 'nova.netconf')
 CONF.import_opt('my_ip', 'nova.netconf')
 CONF.import_opt('vnc_enabled', 'nova.vnc')
 CONF.import_opt('enabled', 'nova.spice', group='spice')
+CONF.import_opt('enable', 'nova.cells.opts', group='cells')
 
 LOG = logging.getLogger(__name__)
 
@@ -340,6 +342,7 @@ class ComputeManager(manager.SchedulerDependentManager):
         self.is_quantum_security_groups = (
             openstack_driver.is_quantum_security_groups())
         self.consoleauth_rpcapi = consoleauth.rpcapi.ConsoleAuthAPI()
+        self.cells_rpcapi = cells_rpcapi.CellsAPI()
 
         super(ComputeManager, self).__init__(service_name="compute",
                                              *args, **kwargs)
@@ -1301,8 +1304,12 @@ class ComputeManager(manager.SchedulerDependentManager):
                 system_metadata=system_meta)
 
         if CONF.vnc_enabled or CONF.spice.enabled:
-            self.consoleauth_rpcapi.delete_tokens_for_instance(context,
-                                                       instance['uuid'])
+            if CONF.cells.enable:
+                self.cells_rpcapi.consoleauth_delete_tokens(context,
+                        instance['uuid'])
+            else:
+                self.consoleauth_rpcapi.delete_tokens_for_instance(context,
+                        instance['uuid'])
 
     @exception.wrap_exception(notifier=notifier, publisher_id=publisher_id())
     @wrap_instance_event
